split compile time and runtime envt, made tail recursive
This is probably rather bad Rust code. It is the first time I ever
did anything practical with rust.

So I wanted to repeat the logical next steps: make it tail recursive
an split compile time code analysis from runtime.

So far things seem to work at least. Except for "set!" which should not
be too hard and comming one day.

Speed is disappointing.

Some code cleanup is still to be done.

Open Questions:

1) I wonder: likely I should understand when to use Rc vs. Box; I
guess there is useless copying going on.

2) The Op::PRINTLN is a work around to just make it work some way. It
should be exported as NativeFn. For this I would need to create a
closure and define it. Grep for "FIXME: can't get that to work
properly" to find my failed attempts.

3) I would actually rather like to be able to declare a version on
NativeFn which is FnMut(). Can't figure out how. Any help appreciated.